#14100: add common function and delete old strided slice code
